backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2